% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_ID_Num.R
\name{agCount_ID_Num}
\alias{agCount_ID_Num}
\title{Count of concept relations using URI/ID via SPARQL.}
\usage{
agCount_ID_Num(
  ID_Name,
  EndPoint,
  FROM = "",
  Property,
  Message = TRUE,
  DirSave = FALSE,
  Dir = "R01_Results"
)
}
\arguments{
\item{ID_Name}{a character vector corresponing to the entity ID.}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{Property}{a list of two character vectors.
The first element of list contain the first property ID.
The second element of list contain the second property ID.}

\item{Message}{logical; perform an output of EntityName or not.}

\item{DirSave}{logical; save the results in the Dir path or not}

\item{Dir}{a folder path for output files.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using an entity URI via SPARQL.
the specific functions for each endpoint were also prepared.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
ID <- "wd:Q81163"

print(KzLabEndPoint)
print(wikidataClassProperty)

#run
agCount_ID_Num(
  ID_Name=ID,
  EndPoint=KzLabEndPoint$EndPoint,
  FROM=KzLabEndPoint$FROM,
  Property=wikidataClassProperty)

}
}
\author{
Satoshi Kume
}
