% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_ID_Num_Wikidata_QID_P279_P31.R
\name{agCount_ID_Num_Wikidata_QID_P279_P31}
\alias{agCount_ID_Num_Wikidata_QID_P279_P31}
\title{Count of concept relations in wikidata using URI/ID via SPARQL.}
\usage{
agCount_ID_Num_Wikidata_QID_P279_P31(ID_Name, Message = FALSE)
}
\arguments{
\item{ID_Name}{a character vector corresponing to the entity ID.}

\item{Message}{logical; perform an output of EntityName or not.}
}
\value{
data.frame
}
\description{
this function count of concept relations of subClassOf (wdt:P279)
and instanceOf (wdt:P31) in wikidata RDF data using an entity URI via SPARQL.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
ID <- "wd:Q81163"

agCount_ID_Num_Wikidata_QID_P279_P31(
  ID_Name=ID,
  Message=TRUE)

#Parallel processing of 4 cores using furrr package
library(furrr)
plan(multisession(workers = 4))
#plan()

#Run multisession
IDs <- c("wd:Q81163", "wd:Q422649", "wd:Q1241898", "wd:Q706")

future_map(as.character(unlist(IDs)), agCount_ID_Num_Wikidata_QID_P279_P31, .progress = TRUE)

}
}
\author{
Satoshi Kume
}
