% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_ID_Prop_Obj_Wikidata_vP.R
\name{agCount_ID_Prop_Obj_Wikidata_vP}
\alias{agCount_ID_Prop_Obj_Wikidata_vP}
\title{Counting the particular triples w/wo GroupBy Option in WIkidata.}
\usage{
agCount_ID_Prop_Obj_Wikidata_vP(Entity_ID, Object, Val = TRUE)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Object}{a vector of character; a string, a value, an entity URI / prefixed ID, or a variable starting with ?.}

\item{Val}{return the results as values or TRUE/FALSE.
If TRUE, this gives a value to hit in the data.
If FALSE, this gives a logical to hit in the data or not.}
}
\value{
data.frame
}
\description{
This function is a function for users.
This is to count the triples with the variable of ?p in wikidata.
}
\examples{
\dontrun{

#Tanaka (Q12159869)
ID <- "wd:Q12159869"
#family name (Q101352)
Obj <- "wd:Q101352"

agCount_ID_Prop_Obj_Wikidata_vP(
  Entity_ID=ID,
  Object=Obj
  )

#show the SPARQL query
CkeckQuery_agCount_ID_Prop_Obj_Wikidata_vP( ID, Obj )

#Parallel processing of two variables and 4 cores using furrr package
library(furrr)
plan(multisession(workers = 4))
#plan()

#Run multisession
IDs <- c("wd:Q81163", "wd:Q422649", "wd:Q1241898", "wd:Q706")
Objs <- c("wd:Q101352", "wd:Q101352", "wd:Q101352", "wd:Q101352")
furrr::future_map2(IDs, Objs, agCount_ID_Prop_Obj_Wikidata_vP, .progress = TRUE)

}
}
\author{
Satoshi Kume
}
