% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_ID_Prop_Obj_Wikidata_vP_vO.R
\name{agCount_ID_Prop_Obj_Wikidata_vP_vO}
\alias{agCount_ID_Prop_Obj_Wikidata_vP_vO}
\title{Counting the particular triples w/wo GroupBy Option in WIkidata.}
\usage{
agCount_ID_Prop_Obj_Wikidata_vP_vO(Entity_ID, Val = TRUE)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Val}{return the results as values or TRUE/FALSE.
If TRUE, this gives a value to hit in the data.
If FALSE, this gives a logical to hit in the data or not.}
}
\value{
data.frame
}
\description{
This function is a function for users.
This is to count the triples with the variable of ?p in wikidata.
}
\examples{
\dontrun{

ID <- "wd:P31"
agCount_ID_Prop_Obj_Wikidata_vP_vO( Entity_ID=ID, Val=TRUE )

#show the SPARQL query
CkeckQuery_agCount_ID_Prop_Obj_Wikidata_vP_vO( ID )

#Parallel processing of two variables and 4 cores using furrr package
library(furrr)
plan(multisession(workers = 4))
#plan()

#Run multisession
IDs <- c("wd:Q81163", "wd:Q422649", "wd:Q1241898", "wd:Q706")
furrr::future_map(IDs, agCount_ID_Prop_Obj_Wikidata_vP_vO, .progress = TRUE)

}
}
\author{
Satoshi Kume
}
