% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_QID_Property_Object.R,
%   R/agCount_QID_Property_Object_Wikidata.R
\name{agCount_ID_Property_Object}
\alias{agCount_ID_Property_Object}
\title{Counting the particular triples w/wo GroupBy Option.}
\usage{
agCount_ID_Property_Object(
  Entity_ID,
  Property,
  Object,
  Count,
  GroupBy = FALSE,
  EndPoint,
  FROM,
  Message = TRUE
)

agCount_ID_Property_Object(
  Entity_ID,
  Property,
  Object,
  Count,
  GroupBy = FALSE,
  EndPoint,
  FROM,
  Message = TRUE
)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Property}{a vector of character; a property URI / prefixed ID, or a variable starting with ?.}

\item{Object}{a vector of character; a string, a value, an entity URI / prefixed ID, or a variable starting with ?.}

\item{Count}{a vector of character; the corresponding variable out of Entity_ID, Property, or Object.}

\item{GroupBy}{logical; default is FALSE.
if TRUE, results were sorted by the variable of Count.}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{Message}{logical; perform an output of Entity_ID or not.

 @description This function counts the particular triples w/wo GroupBy Option.
This is the flexible function for the developer, and then
the users should use child function for counting the particular triples.}

\item{A}{}
}
\value{
data.frame

data.frame
}
\description{
A
}
\examples{
\dontrun{

ID <- "wd:Q12159869"
Prop <- "?p"
Obj <- "wd:Q101352"
Count <- "?p"

EndPoint <- KzLabEndPoint_Wikidata$EndPoint
FROM <- KzLabEndPoint_Wikidata$FROM

agCount_ID_Property_Object(
  Entity_ID=ID,
  Property=Prop,
  Object=Obj,
  Count=Prop,
  GroupBy=FALSE,
  EndPoint=EndPoint,
  FROM=FROM
  )

}
\dontrun{

ID <- "wd:Q12159869"
Prop <- "?p"
Obj <- "wd:Q101352"
Count <- "?p"

EndPoint <- KzLabEndPoint_Wikidata$EndPoint
FROM <- KzLabEndPoint_Wikidata$FROM

agCount_ID_Property_Object(
  Entity_ID=ID,
  Property=Prop,
  Object=Obj,
  Count=Prop,
  GroupBy=FALSE,
  EndPoint=EndPoint,
  FROM=FROM
  )

}
}
\author{
Satoshi Kume

Satoshi Kume
}
