% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_Label_Num.R
\name{agCount_Label_Num}
\alias{agCount_Label_Num}
\title{Count of labels and concept relations using labels via SPARQL.}
\usage{
agCount_Label_Num(
  Entity_Name,
  EndPoint,
  FROM = "",
  Property,
  Message = FALSE,
  FilterRegex = FALSE,
  DirSave = FALSE,
  Dir = "R01_Results"
)
}
\arguments{
\item{Entity_Name}{a character vector. The string was
automatically judged to be Japanese (@ja) or English (@en)}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{Property}{a list of two character vectors.
The first element of list contain the first property ID.
The second element of list contain the second property ID.}

\item{Message}{logical; perform an output of Entity_Name or not.}

\item{FilterRegex}{do not use this option.}

\item{DirSave}{logical; save the results in the Dir path or not}

\item{Dir}{a folder path for output files.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using a label information via SPARQL.
the specific functions for each endpoint were also prepared.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
Label <- "polymer"

print(KzLabEndPoint)
print(wikidataClassProperty)

#run
agCount_Label_Num(
  Entity_Name=Label,
  EndPoint=KzLabEndPoint$EndPoint,
  FROM=KzLabEndPoint$FROM,
  Property=wikidataClassProperty)
}
}
\author{
Satoshi Kume
}
