% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_Label_Num_MeshRDF.R
\name{agCount_Label_Num_MeshRDF}
\alias{agCount_Label_Num_MeshRDF}
\title{Count triples from a label query via SPARQL for Mesh RDF.}
\usage{
agCount_Label_Num_MeshRDF(Entity_Name, Dir, FilterRegex = TRUE)
}
\arguments{
\item{Entity_Name}{a character vector.
The string should be in English (@en)}

\item{Dir}{a folder path for output files.}
}
\value{
data.frame
}
\description{
This function allow to count labels and class relations for
Mesh RDF endpoint via SPARQL.
For parameters of EndPoint and graph id, the variable of EndPoint_MeshRDF is used.
For parameters of properties, the variable of MeshRDF_ClassProperty is used.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
Entity_Name <- "Polymers"
Dir <- "Mesh_test"

#run SPARQL
agCount_Label_Num_MeshRDF(
  Entity_Name=Entity_Name,
  Dir=Dir
  )

#show the SPARQL query
CkeckQuery_agCount_Label_Num_MeshRDF(
  Entity_Name=Entity_Name,
  Dir=Dir
  )

#Parallel processing of 4 cores using furrr package
library(furrr)
plan(multisession(workers = 4))
#plan()

#prepare a vector of labels
Labs <- c("polymer", "protein", "Pteridophyta", "material substance")

#run multisession
results <- future_map(Labs, agCount_Label_Num_MeshRDF, .progress=T)
results

}

}
\author{
Satoshi Kume
}
