% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_QID_Property_Object_Wikidata.R
\name{agCount_QID_Property_Object_Wikidata_vP}
\alias{agCount_QID_Property_Object_Wikidata_vP}
\title{Counting the particular triples w/wo GroupBy Option in WIkidata.}
\usage{
agCount_QID_Property_Object_Wikidata_vP(Entity_ID, Object)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Object}{a vector of character; a string, a value, an entity URI / prefixed ID, or a variable starting with ?.}
}
\value{
data.frame
}
\description{
This function is a function for users.
This is to count the triples with the variable of ?p in wikidata.
}
\examples{
\dontrun{

ID <- "wd:Q12159869"
Obj <- "wd:Q101352"

agCount_QID_Property_Object_Wikidata_vP(
  Entity_ID=ID,
  Object=Obj,
  )

#Parallel processing of two variables and 4 cores using furrr package
library(furrr)
plan(multisession(workers = 4))
#plan()

#Run multisession
IDs <- c("wd:Q81163", "wd:Q422649", "wd:Q1241898", "wd:Q706")

furrr::future_map(as.character(unlist(IDs)), agCount_ID_Num_Wikidata_QID_P279_P31, .progress = TRUE)

}
}
\author{
Satoshi Kume
}
