\name{agGraphSearch-package}
\alias{agGraphSearch-package}
\alias{agGraphSearch}
\docType{package}
\title{
\packageTitle{agGraphSearch}
}
\description{
\packageDescription{agGraphSearch}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{agGraphSearch}
\packageIndices{agGraphSearch}
}
\author{
\packageAuthor{agGraphSearch}

Maintainer: \packageMaintainer{agGraphSearch}
}

\keyword{ package }
\examples{
ls("package:agGraphSearch") # This is empty package
}