% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agIDtoLabel.R
\name{agIDtoLabel}
\alias{agIDtoLabel}
\title{Searching the label corresponding to the entity url via SPARQL.}
\usage{
agIDtoLabel(
  Entity_ID,
  EndPoint,
  FROM,
  PropertyForLabel = 1,
  lang = 1,
  Message = TRUE
)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{PropertyForLabel}{a numeric vector.
if the number is 1 (default), the label relation use "rdfs:label".
if the number is 2, the label relation use " skos:prefLabel".}

\item{lang}{a numeric vector.
if the number is 1 (default), the label contains English label (@en).
if the number is 2, the label contains Japanese label (@ja).
if the number is 3, the labels contain both English and Japanese labels.}

\item{Message}{logical; perform an output of Entity_ID or not.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using an entity URI via SPARQL.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
ID <- "wd:Q81163"

print(KzLabEndPoint)
print(wikidataClassProperty)

#run
agIDtoLabel(
  Entity_ID=ID,
  EndPoint=KzLabEndPoint$EndPoint,
  FROM=KzLabEndPoint$FROM
  )

}
}
\author{
Satoshi Kume
}
