% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agJsTreeR.R
\name{agJsTreeR}
\alias{agJsTreeR}
\title{agJsTreeR: jsTree for agGraphSearch (modified)}
\usage{
agJsTreeR(
  Data,
  TopClass,
  Brouse = TRUE,
  SaveHtmlFile = TRUE,
  FileName = NULL,
  Output = NULL,
  ExcludeQID = TRUE
)
}
\arguments{
\item{Data}{a input data.frame.}

\item{TopClass}{a string of the top class. ex. entity.wd:Q35120.}

\item{Brouse}{a logical}

\item{SaveHtmlFile}{a logical}

\item{FileName}{a character of file name}

\item{Output}{a character of folder directory}

\item{ExcludeQID}{a logical}
}
\value{
html
}
\description{
this function is to use jsTree for agGraphSearch data.
}
\author{
Satoshi Kume
}
