% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agQIDtoLabel_Wikidata.R
\name{agQIDtoLabel_Wikidata}
\alias{agQIDtoLabel_Wikidata}
\title{Searching the label corresponding to the entity url in wikidata via SPARQL.}
\usage{
agQIDtoLabel_Wikidata(Entity_ID, lang = 3, Message = TRUE)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}

\item{lang}{a numeric vector.
if the number is 1, the label contains English label (@en).
if the number is 2, the label contains Japanese label (@ja).
if the number is 3 (default), the labels contain both English and Japanese labels.}

\item{Message}{logical; perform an output of EntityName or not.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using an entity URI via SPARQL.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
ID <- "wd:Q81163"

#run
agQIDtoLabel_Wikidata( Entity_ID = ID )

}
}
\author{
Satoshi Kume
}
