% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTableDT.R
\name{agTableDT}
\alias{agTableDT}
\title{Interactive html table using DT package.}
\usage{
agTableDT(
  Data,
  Output = TRUE,
  File = paste0("agTableDT_table_", format(Sys.time(), "\%y\%m\%d_\%H\%M"), ".html"),
  Rownames = FALSE,
  AutoWidth = TRUE,
  Browse = TRUE,
  Editable = FALSE,
  Caption = "Table. results",
  Width = "300px",
  PageLength = 25
)
}
\arguments{
\item{Data}{an input as a data.frame.}

\item{Output}{logical; if TRUE, create the output file as a name of File.}

\item{File}{a string of file name.}

\item{Rownames}{logical.}

\item{AutoWidth}{logical.}

\item{Browse}{logical.}

\item{Editable}{logical.}

\item{Caption}{a string of caption.}

\item{Width}{a string of px; ex. '300px'.}

\item{PageLength}{a numeric vector.}
}
\value{
DT::datatable
}
\description{
this function creates a interactive table for checking the data.
}
\examples{
\dontrun{

print(agGraphSearch::URI2Prefix)

#run
agTableDT( Data = agGraphSearch::URI2Prefix, PageLength = 10 )

}
}
\author{
Satoshi Kume
}
