% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agTableKB.R
\name{agTableKB}
\alias{agTableKB}
\title{Text table using knitr package.}
\usage{
agTableKB(Data, Head = TRUE, Format = 1, Align = 1)
}
\arguments{
\item{Data}{an input as a data.frame.}

\item{Head}{logical; if TRUE, show the header only.}

\item{Format}{a numeric vector.
if the value is 1, the output format is "pipe".
if the value is 2, the output format is "rst".
if the value is 3, the output format is "simple".
if the value is 4, the output format is "latex".}

\item{Align}{a numeric vector.
if the value is 1, the output format is "aligned on center".
if the value is 2, the output format is "aligned on left".
if the value is 3, the output format is "aligned on right".
if the value is 4, the output format is "NULL".}
}
\value{
knitr_kable
}
\description{
this function creates a text table for checking the data.
}
\examples{
\dontrun{

head(iris)

#run
agTableKB( Data = iris )

}
}
\author{
Satoshi Kume
}
