% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_Alt_Wikidata.R
\name{agWD_Alt_Wikidata}
\alias{agWD_Alt_Wikidata}
\title{Searching the labels in the wikidata endopoint via SPARQL.}
\usage{
agWD_Alt_Wikidata(Entity_Name)
}
\arguments{
\item{Entity_Name}{a character vector corresponing to the entity label.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using an entity URI via SPARQL.
}
\examples{
\dontrun{

#CAS Registry Number (wikidata prefix URI: wd:Q102507)
Label <- "CAS Registry Number"

#view the SPARQL query
CkeckQuery_agWD_Alt_Wikidata( Entity_Name=Label )

#run SPARQL
agWD_Alt_Wikidata( Entity_Name=Label )

}
}
\author{
Satoshi Kume
}
