% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_ID_prop_Obj.R
\name{agWD_ID_prop_Obj}
\alias{agWD_ID_prop_Obj}
\title{Counting the particular triples w/wo GroupBy Option.}
\usage{
agWD_ID_prop_Obj(
  Entity_ID,
  Property,
  Object,
  DISTINCT = TRUE,
  PREFIX_uri = TRUE,
  lang = 1,
  EndPoint,
  FROM,
  Message = TRUE
)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Property}{a vector of character; a property URI / prefixed ID, or a variable starting with ?.}

\item{Object}{a vector of character; a string, a value, an entity URI / prefixed ID, or a variable starting with ?.}

\item{PREFIX_uri}{}

\item{lang}{a numeric vector.
if the number is 1 (default), the label contains English label (@en).
if the number is 2, the label contains Japanese label (@ja).
if the number is 3, the labels contain both English and Japanese labels.}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{Message}{logical; perform an output of Entity_ID or not.}
}
\value{
data.frame
}
\description{
This function counts the particular triples w/wo GroupBy Option.
This is the flexible function for the developer, and then
the users should use child function for counting the particular triples.
}
\examples{
\dontrun{

#Tanaka (Q12159869)
ID <- "wd:Q12159869"

#variables
Prop <- "wdt:P31"
Obj <- "?o"

EndPoint <- KzLabEndPoint_Wikidata$EndPoint
FROM <- KzLabEndPoint_Wikidata$FROM

agWD_ID_prop_Obj(
  Entity_ID=ID,
  Property=Prop,
  Object=Obj,
  DISTINCT=TRUE,
  PREFIX=TRUE,
  EndPoint=EndPoint,
  FROM=FROM
  )

}
}
\author{
Satoshi Kume
}
