% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_ID_Prop_Obj_Wikidata_vO.R
\name{agWD_ID_Prop_Obj_Wikidata_vO}
\alias{agWD_ID_Prop_Obj_Wikidata_vO}
\title{Counting the particular triples w/wo GroupBy Option.}
\usage{
agWD_ID_Prop_Obj_Wikidata_vO(Entity_ID, Property, lang = 3)
}
\arguments{
\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Property}{a vector of character; a property URI / prefixed ID, or a variable starting with ?.}

\item{lang}{a numeric vector.
if the number is 1, the label contains English label (@en).
if the number is 2, the label contains Japanese label (@ja).
if the number is 3 (default), the labels contain both English and Japanese labels.}
}
\value{
data.frame
}
\description{
This function counts the particular triples w/wo GroupBy Option.
This is the flexible function for the developer, and then
the users should use child function for counting the particular triples.
}
\examples{
\dontrun{

#Tanaka (Q12159869)
ID <- "wd:Q12159869"

#variables
Prop <- "wdt:P31"

agWD_ID_Prop_Obj_Wikidata_vO(
  Entity_ID=ID,
  Property=Prop
  )

}
}
\author{
Satoshi Kume
}
