% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_ID_prop_Obj_Wikidata_P31.R
\name{agWD_QID_property_Object_v01}
\alias{agWD_QID_property_Object_v01}
\title{Counting the particular triples w/wo GroupBy Option.}
\usage{
agWD_QID_property_Object_v01(
  WDname = "wd:Q12159869",
  Property = "wdt:P31",
  Object = "?o",
  DISTINCT = T,
  PREFIX = T,
  FROM = "From <http://wikidata_nearly_full_201127> ",
  EndPoint = "http://kozaki-lab.osakac.ac.jp/agraph/NEDO_pj"
)
}
\arguments{
\item{Property}{a vector of character; a property URI / prefixed ID, or a variable starting with ?.}

\item{Object}{a vector of character; a string, a value, an entity URI / prefixed ID, or a variable starting with ?.}

\item{FROM}{a string of graph URI in the endpoint. The default is blank ("").}

\item{EndPoint}{a string of SPARQL endpoint. ex. http://....}

\item{Entity_ID}{a vector of character; an entity URI / prefixed ID, or a variable starting with ?.}

\item{Count}{a vector of character; the corresponding variable out of Entity_ID, Property, or Object.}

\item{GroupBy}{logical; default is FALSE.
if TRUE, results were sorted by the variable of Count.}

\item{Message}{logical; perform an output of Entity_ID or not.

 @description This function counts the particular triples w/wo GroupBy Option.
This is the flexible function for the developer, and then
the users should use child function for counting the particular triples.}
}
\value{
data.frame
}
\description{
Counting the particular triples w/wo GroupBy Option.
}
\examples{
\dontrun{


#Tanaka (Q12159869)
ID <- "wd:Q12159869"
#family name (Q101352)
Obj <- "wd:Q101352"
Prop <- "?p"
Count <- "?p"

EndPoint <- KzLabEndPoint_Wikidata$EndPoint
FROM <- KzLabEndPoint_Wikidata$FROM

agCount_ID_Prop_Obj(
  Entity_ID=ID,
  Property=Prop,
  Object=Obj,
  Count=Prop,
  GroupBy=FALSE,
  EndPoint=EndPoint,
  FROM=FROM
  )

}
}
\author{
Satoshi Kume
}
