% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_ID_Num_Wikidata_QID_P279_P31.R
\name{CkeckQuery_agCount_ID_Num_Wikidata_QID_P279_P31}
\alias{CkeckQuery_agCount_ID_Num_Wikidata_QID_P279_P31}
\title{View the SPARQL query of agCount_ID_Num_Wikidata_QID_P279_P31.}
\usage{
CkeckQuery_agCount_ID_Num_Wikidata_QID_P279_P31(Entity_ID)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}
}
\value{
message
}
\description{
View the SPARQL query of agCount_ID_Num_Wikidata_QID_P279_P31.
}
\seealso{
agCount_ID_Num_Wikidata_QID_P279_P31
}
\author{
Satoshi Kume
}
