% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_Label_Num_MeshRDF.R
\name{CkeckQuery_agCount_Label_Num_MeshRDF}
\alias{CkeckQuery_agCount_Label_Num_MeshRDF}
\title{View the SPARQL query of agCount_Label_Num_MeshRDF.}
\usage{
CkeckQuery_agCount_Label_Num_MeshRDF(Entity_Name)
}
\arguments{
\item{Entity_Name}{a character vector.
The string should be in English (@en).}
}
\value{
message
}
\description{
View the SPARQL query of agCount_Label_Num_MeshRDF.
}
\examples{
\dontrun{

CkeckQuery_agCount_Label_Num_MeshRDF( "Polymers" )

}

}
\seealso{
agCount_Label_Num_MeshRDF
}
\author{
Satoshi Kume
}
