% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_Label_Num_Wikidata_P279_P31.R
\name{CkeckQuery_agCount_Label_Num_Wikidata_P279_P31}
\alias{CkeckQuery_agCount_Label_Num_Wikidata_P279_P31}
\title{View the SPARQL query of agCount_Label_Num_Wikidata_P279_P31.}
\usage{
CkeckQuery_agCount_Label_Num_Wikidata_P279_P31(Entity_Name)
}
\arguments{
\item{Entity_Name}{a character vector. The string was
automatically judged to be Japanese (@ja) or English (@en)}
}
\value{
message
}
\description{
View the SPARQL query of agCount_Label_Num_Wikidata_P279_P31.
}
\seealso{
agCount_Label_Num_Wikidata_P279_P31
}
\author{
Satoshi Kume
}
