% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agQIDtoLabel_Wikidata.R
\name{CkeckQuery_agQIDtoLabel_Wikidata}
\alias{CkeckQuery_agQIDtoLabel_Wikidata}
\title{View the SPARQL query of agQIDtoLabel_Wikidata}
\usage{
CkeckQuery_agQIDtoLabel_Wikidata(Entity_ID, lang = 3)
}
\arguments{
\item{lang}{a numeric vector.
if the number is 1, the label contains English label (@en).
if the number is 2, the label contains Japanese label (@ja).
if the number is 3 (default), the labels contain both English and Japanese labels.}

\item{Entity_Name}{a character vector. The string was
automatically judged to be Japanese (@ja) or English (@en)}
}
\value{
message
}
\description{
View the SPARQL query of agQIDtoLabel_Wikidata
}
\seealso{
agCount_Label_Num_Wikidata_P279_P31
}
\author{
Satoshi Kume
}
