% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_Alt_Wikidata.R
\name{CkeckQuery_agWD_Alt_Wikidata}
\alias{CkeckQuery_agWD_Alt_Wikidata}
\title{View SPARQ Query of agWD_Alt_Wikidata.}
\usage{
CkeckQuery_agWD_Alt_Wikidata(Entity_Name)
}
\arguments{
\item{Entity_Name}{a character vector corresponing to the entity label.}
}
\description{
View SPARQ Query of agWD_Alt_Wikidata.
}
\seealso{
agWD_Alt_Wikidata
}
