% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeDistance.R
\name{ComputeDistance}
\alias{ComputeDistance}
\title{Compute distance metrics between input strings and rdfs}
\usage{
ComputeDistance(InputTerms, RDFterms, q = 2, nthread = 4, TopWords = 3)
}
\arguments{
\item{InputTerms}{a vector of character; a target list of words.}

\item{RDFterms}{a vector of character.}

\item{q}{Size of the q-gram}

\item{nthread}{Maximum number of threads to use.}

\item{TopWords}{Number of words with top strings distances}
}
\value{
data.frame
}
\description{
This function counts the particular relations between strings.
Perfect Match, partial Match, strings similarity using
cosine distance between q-gram profiles (cosine, N=2).
}
\examples{
\dontrun{

library(magrittr)

InputTerms <- "polymer"
RDFterms <- "Polymers"

ComputeDistance(
  InputTerms,
  RDFterms
  )
}

}
\author{
Satoshi Kume
}
