% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\docType{data}
\name{KzLabEndPoint_Wikidata}
\alias{KzLabEndPoint_Wikidata}
\title{Default Parameters}
\format{
An object of class \code{list} of length 2.
}
\usage{
KzLabEndPoint_Wikidata
}
\description{
Default Parameters
}
\keyword{datasets}
