% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Label2Jpn.R
\name{Label2Jpn}
\alias{Label2Jpn}
\title{Translate labels to Japanese}
\usage{
Label2Jpn(Data, Auth_Key = "Your Authentication Key")
}
\arguments{
\item{Data}{an input data}

\item{Auth_Key}{a string, Your Authentication Key}
}
\value{
data.frame
}
\description{
Translate labels to Japanese
}
\author{
Satoshi Kume
}
