% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcR.R
\name{ListDF2DF}
\alias{ListDF2DF}
\title{Bind list to Data Frame}
\usage{
ListDF2DF(input)
}
\arguments{
\item{input}{A list of data frames. their colnames should be same.}
}
\value{
data.frame
}
\description{
Create a data frame from a list of multiple data frames with the same colnames.
}
\examples{
\dontrun{
#parameters

#test dat
test <- list(data.frame(x=c(1:3), y=c(1:3)),
             data.frame(x=c(4:6), y=c(4:6)),
             data.frame(x=c(7:20), y=c(7:20)))

#run
ListDF2DF( input=test )

}
}
\author{
Satoshi Kume
}
