% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeChEBIData.R
\name{MergeChEBIData}
\alias{MergeChEBIData}
\title{Merge ChEBI Labels}
\usage{
MergeChEBIData(Data, Labels)
}
\arguments{
\item{Data}{a RDF data.}

\item{Labels}{a RDF data.}
}
\value{
data.frame
}
\description{
This function is ...
}
\author{
Satoshi Kume
}
