% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeMeshData.R
\name{MergeMeshData}
\alias{MergeMeshData}
\title{Merge Labels}
\usage{
MergeMeshData(Data, Labels)
}
\arguments{
\item{Data}{a RDF data.}

\item{Labels}{a RDF data.}
}
\value{
data.frame
}
\description{
This function is ...
}
\author{
Satoshi Kume
}
