% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeWikiData.R
\name{MergeWikiData}
\alias{MergeWikiData}
\title{Merge Labels}
\usage{
MergeWikiData(Data, Labels)
}
\arguments{
\item{Data}{a RDF data.}

\item{Labels}{a RDF data.}
}
\value{
data.frame
}
\description{
This function is ...
}
\author{
Satoshi Kume
}
