% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste1.R
\name{paste1}
\alias{paste1}
\title{File name}
\usage{
paste1(FirstWords, extension = "html")
}
\arguments{
\item{FirstWords}{a first string of output.}

\item{extension}{a character of file extension.}
}
\value{
character
}
\description{
File name
}
\author{
Satoshi Kume
}
