% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_plot.R
\name{pc_plot}
\alias{pc_plot}
\title{Visualize the results for common upper-level entities}
\usage{
pc_plot(input, SaveFolder, FileName, IDnum)
}
\value{
data.frame
}
\description{
Visualize the results for common upper-level entities
}
\examples{
\dontrun{
}
}
\author{
Satoshi Kume
}
