% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyPath.R
\name{PropertyPath}
\alias{PropertyPath}
\title{Search class-related hierarchy by the property path of SPARQL query.}
\usage{
PropertyPath(
  Entity_ID,
  PropertyPathDepth,
  Upper = TRUE,
  p1,
  p2,
  Message = FALSE,
  JA = TRUE,
  PREFIX_uri = TRUE,
  FROM,
  EndPoint,
  FolferPath
)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}

\item{Message}{logical; perform an output of EntityName or not.}
}
\description{
Search class-related hierarchy by the property path of SPARQL query.
}
