% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyPath_GraphUp_Wikidata.R
\name{PropertyPath_GraphUp_Wikidata}
\alias{PropertyPath_GraphUp_Wikidata}
\title{Search class-related hierarchy in wikidata by the property path of SPARQL query.}
\usage{
PropertyPath_GraphUp_Wikidata(Entity_ID, Depth = 30, Message = FALSE)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}

\item{Message}{logical; perform an output of EntityName or not.}
}
\description{
Search class-related hierarchy in wikidata by the property path of SPARQL query.
}
