% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{ProxySet}
\alias{ProxySet}
\title{Proxy settings}
\usage{
ProxySet(proxy_url)
}
\arguments{
\item{proxy_url}{a character vector of proxy URI.}
}
\description{
This function provides the proxy settings.
}
\examples{
\dontrun{

proxy_url = "http://wwwproxy.osakac.ac.jp:8080"
ProxySet(proxy_url)

}

}
\author{
Satoshi Kume
}
