% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove_vals.R
\name{Remove_vals}
\alias{Remove_vals}
\title{Remove_vals: remove unnecessary strings for Japanese.}
\usage{
Remove_vals(Data, StringLength = 30)
}
\arguments{
\item{Data}{a data frame with columns of "Subject", "Property", "Object", and "OtherInfo".}

\item{StringLength}{a numeric; upper limit of string length.}
}
\value{
data.frame
}
\description{
Exclude one character, number rules and symbols.
}
\author{
Satoshi Kume
}
