% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiQID2Label.R
\name{wikiQID2Label}
\alias{wikiQID2Label}
\title{convert QID to Label via SPARQL query}
\usage{
wikiQID2Label(EntityQID, lang = 1, Message = F)
}
\arguments{
\item{EntityQID}{a character of QID with wd:.}

\item{lang}{a numeric: 1: jpn, 2, en.}

\item{Message}{a logical}
}
\value{
data.frame
}
\description{
This function convert QID to Label via SPARQL query.
}
\author{
Satoshi Kume
}
