% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agCount_Label_Num_Wikidata_P279_P31.R
\name{agCount_Label_Num_Wikidata_P279_P31}
\alias{agCount_Label_Num_Wikidata_P279_P31}
\title{Count of labels and class relations for wikidata local endpoint at OECU via SPARQL.}
\usage{
agCount_Label_Num_Wikidata_P279_P31(EntityName, Message = TRUE)
}
\arguments{
\item{EntityName}{a character vector. The string was
automatically judged to be Japanese (@ja) or English (@en)}

\item{Message}{logical; perform an output of EntityName or not.}
}
\value{
data.frame
}
\description{
For parameters of EndPoint and graph id, the variable of KzLabEndPoint is used.
For parameters of properties, the variable of wikidataClassProperty is used.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
Label <- "polymer"

print(KzLabEndPoint)
print(wikidataClassProperty)

#run
agCount_Label_Num_Wikidata_P279_P31(
  EntityName=Label,
  Message=FALSE
  )
}
}
\author{
Satoshi Kume
}
