% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agGraphAnalysis.R
\name{agGraphAnalysis}
\alias{agGraphAnalysis}
\title{Graph analysis}
\usage{
agGraphAnalysis(
  graphList,
  list1,
  list2,
  UpdateList = F,
  breakRepeat = 20,
  View = T,
  RemoveGraph = F,
  PlusLabel = T,
  FileName = T,
  FileName00,
  OutputResults = T,
  ResultsEach = F,
  LowerSearch = F,
  Entity_only = T,
  list2.WD = "wd:Q35120",
  GraphView01 = F,
  GraphView02 = F,
  GraphView03 = T,
  GraphView04 = T,
  LvView = T,
  WindowSize01 = 10,
  WindowSize02 = 10,
  WindowSize03 = 10,
  WindowSize04 = 7.5,
  WindowSize05 = 7.5
)
}
\arguments{
\item{graphList}{}
}
\value{
data.frame
}
\description{
this function is a general function for
}
\examples{
\dontrun{


}
}
\author{
Satoshi Kume
}
