% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agGraphAnalysis.R
\name{agGraphAnalysis}
\alias{agGraphAnalysis}
\title{Graph analysis function for the upper-level concepts.}
\usage{
agGraphAnalysis(
  graphList,
  search_List,
  CU_List,
  PlusLabel = TRUE,
  ParticularEntity_only = TRUE,
  top.WD = "wd:Q35120",
  UpdateList = FALSE,
  breakRepeat = 20,
  View = TRUE,
  RemoveGraph = FALSE,
  FileName = TRUE,
  OutputResults = TRUE,
  ResultsEach = FALSE,
  LowerSearch = FALSE,
  LvView = TRUE
)
}
\arguments{
\item{graphList}{a graph list}

\item{search_List}{a character vector of search entities.}

\item{CU_List}{a character vector of common upper-level entities}

\item{PlusLabel}{logical; add the labels or not when createing the network.}

\item{ParticularEntity_only}{a logical}

\item{top.WD}{a character vector}

\item{UpdateList}{a logical}

\item{breakRepeat}{a numeric}

\item{View}{a logical}

\item{RemoveGraph}{a logical}

\item{FileName}{a string}

\item{OutputResults}{a logical}

\item{ResultsEach}{a logical}

\item{LowerSearch}{a logical}

\item{LvView}{a logical}
}
\value{
data.frame
}
\description{
this function is a general function for
determining the exploration Segment
}
\examples{
\dontrun{

library(magrittr)
agGraphAnalysis(graphList=eachGraph,
                search_List=list1a,
                CU_List=list2b)

}

}
\author{
Satoshi Kume
}
