% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agGraphAnalysis.R
\name{agGraphAnalysis}
\alias{agGraphAnalysis}
\title{Graph analysis function for the upper-level concepts.}
\usage{
agGraphAnalysis(
  graphList,
  list1,
  list2,
  ParticularEntity_only = TRUE,
  list2.WD = "wd:Q35120",
  UpdateList = FALSE,
  breakRepeat = 20,
  View = TRUE,
  RemoveGraph = FALSE,
  PlusLabel = TRUE,
  FileName = TRUE,
  FileName00,
  OutputResults = TRUE,
  ResultsEach = FALSE,
  LowerSearch = FALSE,
  GraphView01 = FALSE,
  GraphView02 = FALSE,
  GraphView03 = TRUE,
  GraphView04 = TRUE,
  LvView = TRUE,
  WindowSize01 = 10,
  WindowSize02 = 10,
  WindowSize03 = 10,
  WindowSize04 = 7.5,
  WindowSize05 = 7.5
)
}
\arguments{
\item{list1}{a character vector of search entities.}

\item{list2}{a character vector of common upper-level entities}

\item{PlusLabel}{logical; add the labels or not when createing the network.}
}
\value{
data.frame
}
\description{
this function is a general function for
}
\examples{
\dontrun{


}
}
\author{
Satoshi Kume
}
