% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agVisNetwork.R
\name{agVisNetwork}
\alias{agVisNetwork}
\title{Count triples from a label query via SPARQL.}
\usage{
agVisNetwork(
  Graph,
  NodeColorRandom = FALSE,
  Count = 2,
  Size = 10,
  SmallSize = 5,
  StarSize = 10,
  FontSize = 7,
  HeightSclale = "750px",
  WidthSclale = "110\%",
  SEED = 123,
  Selected = NULL,
  Browse = FALSE,
  Output = FALSE,
  FilePath = paste0("agVisNetwork_", format(Sys.time(), "\%y\%m\%d"), ".html"),
  outputNodesEdges = FALSE
)
}
\arguments{
\item{Graph}{a data.frame}
}
\value{
network diagrams
}
\description{
Count triples from a label query via SPARQL.
}
\author{
Satoshi Kume
}
