% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agWD_Alt_Wikidata.R
\name{agWD_Alt_Wikidata}
\alias{agWD_Alt_Wikidata}
\title{Searching the label corresponding to the entity url via SPARQL.}
\usage{
agWD_Alt_Wikidata(Entity_Name)
}
\arguments{
\item{Entity_Name}{a character vector corresponing to the entity label.}
}
\value{
data.frame
}
\description{
this function is a general function for
searching the RDF data using an entity URI via SPARQL.
}
\examples{
\dontrun{
#parameters

#polymer (wikidata prefix URI: wd:Q81163)
Label <- "CAS Registry Number"

print(KzLabEndPoint)
print(wikidataClassProperty)

#run
agWD_Alt_Wikidata(
  Entity_Name=Label,
  EndPoint=KzLabEndPoint$EndPoint,
  FROM=KzLabEndPoint$FROM
  )

}
}
\author{
Satoshi Kume
}
