% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractUpperConcepts4Mesh.R
\name{extractUpperConcepts4Mesh}
\alias{extractUpperConcepts4Mesh}
\title{The upper class search of Mesh RDF}
\usage{
extractUpperConcepts4Mesh(Lab_List, Data, Labels, broaderProperty)
}
\arguments{
\item{Lab_List}{a character vector of input.}

\item{Data}{a RDF data.}

\item{Labels}{a RDF data.}

\item{broaderProperty}{a string vector of property.}
}
\value{
data.frame
}
\description{
This function is a function to obtain
the upper class (meshv:broaderDescriptor) of Mesh RDF.
}
\author{
Satoshi Kume
}
