% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikiCount_QID_Property_Object.R
\name{wikiCount_QID_Property_Object}
\alias{wikiCount_QID_Property_Object}
\title{wikiCount_QID_Property_Object}
\usage{
wikiCount_QID_Property_Object(
  Entity_ID,
  Object,
  Property = "?p",
  Count = "?p",
  lang = 1,
  GroupBy = FALSE
)
}
\arguments{
\item{Entity_ID}{a character vector corresponing to the entity ID.}

\item{Property}{a character vector}

\item{Count}{a character vector}

\item{lang}{a numeric; The value 1 is to use "ja, en". The value 2 is to use "ja". The value 3 is to use "en".}

\item{GroupBy}{a logical}
}
\value{
data.frame
}
\description{
this function is a general function for ...
}
\author{
Satoshi Kume
}
